xAI: v2 storage rework -- §6.1 CurvSummary + single-slot ladder + zarr cuts (~40x)
Per-run extract footprint was ~800-980 GB (extract store ~80%, dominated by §6.1
block_optim writing the FULL exp_avg_sq/precond spectra = full param numel per layer
per checkpoint, ~2.4 GB/ckpt, with ZERO downstream readers), plus a dense/linear kept
ladder (the latest-keep rule latched every step). All 6 runs did not fit in ~1.8 TB.
This brings per-run to ~15-20 GB (all 6 fit easily).

Core (audit-fleet designed, adversarially verified, smoke-tested):
- block_optim §6.1: replace the full spectra with a fixed-width CurvSummary (top-64
  eigenvalues + 64 log-spaced decay knots + 11 shape scalars: lambda_max/min, trace,
  log-det proxy, participation, entropy effective rank, condition number, anisotropy,
  power-law tail, numel). ~385x cut; every curvature-SHAPE quantity preserved. kron L/R
  spectra + QL/QR eigvecs kept f32. Pilot safety net (PREADAPT_PILOT_FULL_SPECTRA)
  optionally retains raw spectra on a sparse step grid for first-run validation.
- ladder fix (collector/preadapt_common/preadapt_train): on_ladder's exact-power-of-2
  test misses min_step-multiple steps, and the latest-keep rule re-accumulated (re-slim
  only converted FULL->model-only, never deleted). Replace with (a) log2_crossing_rung
  (first emitted step to cross each octave -> stable ~log2 ladder over arbitrary steps),
  (b) a SINGLE rotating kept/latest_full.pt slot (atomic overwrite, step-guarded) as the
  resume anchor -- no pile-up, watch dir drains to 0, (c) resume live->slot->ladder.
- preadapt_common ArrayStore: step-axis chunk batching (reclaim 4K-block inode tax),
  Blosc zstd+shuffle compressor (guarded fallback), dtype= kwarg (default None keeps
  dtype -- f32 guardrail), close()/consolidate_metadata at finalize.
- safe per-block fp16 on bounded arrays (attention stats, probe features w/ CKA n=1024,
  accuracy curves, patch class maps, basin radii). Wide-range spectra/singvals kept f32.

Smoke (.claude_notes_and_utils/testing/smoke_storage_v2.py) passes: codec roundtrips in
reticulate-gpu2, CurvSummary reconstructs lambda_max/trace/decay, log2 ladder = 9 octave
rungs over min_step=12, resume order resolves live->slot->ladder.

+++ xAI/py/extractor/block_basin.py
@@ -287,9 +287,11 @@ def _basin_radius_anisotropy(ctx, model, criterion, images, targets,
 
     # per-direction radii -> array store (the full distribution for post-hoc anisotropy analysis)
     if ctx.array is not None:
-        ctx.array.put(group="basin_radii", step=step, array=arr)
-        ctx.array.put(group="basin_radii_pos", step=step, array=np.asarray(radii_pos, dtype=np.float32))
-        ctx.array.put(group="basin_radii_neg", step=step, array=np.asarray(radii_neg, dtype=np.float32))
+        ctx.array.put(group="basin_radii", step=step, array=arr, dtype=np.float16)
+        ctx.array.put(group="basin_radii_pos", step=step,
+                      array=np.asarray(radii_pos, dtype=np.float32), dtype=np.float16)
+        ctx.array.put(group="basin_radii_neg", step=step,
+                      array=np.asarray(radii_neg, dtype=np.float32), dtype=np.float16)
 
     a64 = arr.astype(np.float64)
     mean = float(np.mean(a64))
